change database Structures ,prepare for yuexian event
