+++ src/grabData/TransientResponse.java
@@ -1,12 +1,12 @@
 package grabData;
 
-import hibernatePOJO.EventTransient;
+import hibernatePOJO.EventPower;
 
 import java.util.List;
 
 public class TransientResponse {
     private String code;
-    private List<EventTransient> result;
+    private List<EventPower> result;
 
     public String getCode() {
         return code;
@@ -16,11 +16,11 @@ public class TransientResponse {
         this.code = code;
     }
 
-    public List<EventTransient> getResult() {
+    public List<EventPower> getResult() {
         return result;
     }
 
-    public void setResult(List<EventTransient> result) {
+    public void setResult(List<EventPower> result) {
         this.result = result;
     }
 }
