+++ src/hibernatePOJO/EventPower.hbm.xml
@@ -5,32 +5,38 @@
 <hibernate-mapping>
 
     <class name="hibernatePOJO.EventPower" table="event_power" schema="test">
-        <id name="epid">
-            <column name="epid" sql-type="int(11)"/>
+        <id name="teid">
+            <column name="teid" sql-type="int(11)"/>
         </id>
         <property name="did">
             <column name="did" sql-type="varchar(255)" not-null="true"/>
         </property>
+        <property name="time">
+            <column name="time" sql-type="datetime"/>
+        </property>
+        <property name="eventtype">
+            <column name="eventtype" sql-type="int(11)"/>
+        </property>
         <property name="type">
-            <column name="type" sql-type="varchar(255)" not-null="true"/>
+            <column name="type" sql-type="int(11)"/>
         </property>
-        <property name="content">
-            <column name="content" sql-type="varchar(255)" not-null="true"/>
+        <property name="subtype">
+            <column name="subtype" sql-type="int(11)"/>
         </property>
-        <property name="occurtime">
-            <column name="occurtime" sql-type="datetime" not-null="true"/>
+        <property name="duration">
+            <column name="duration" sql-type="int(11)"/>
+        </property>
+        <property name="value">
+            <column name="value" sql-type="double" precision="-1"/>
+        </property>
+        <property name="discription">
+            <column name="discription" sql-type="varchar(255)" not-null="true"/>
         </property>
         <property name="signature">
             <column name="signature" sql-type="varchar(255)" not-null="true"/>
         </property>
         <property name="annotation">
             <column name="annotation" sql-type="varchar(255)" not-null="true"/>
-        </property>
-        <property name="etype">
-            <column name="etype" sql-type="varchar(255)" not-null="true"/>
-        </property>
-        <property name="isMark">
-            <column name="isMark" sql-type="int(11)" not-null="true"/>
         </property>
     </class>
 </hibernate-mapping>
