+++ src/hibernatePOJO/EventPower.java
@@ -4,54 +4,88 @@ import java.sql.Timestamp;
 import java.util.Objects;
 
 public class EventPower {
-    private int epid;
+    private int teid;
     private String did;
-    private String type;
-    private String content;
-    private Timestamp occurtime;
+    private Timestamp time;
+    private int eventtype;
+    private int type;
+    private int subtype;
+    private int duration;
+    private double value;
+    private String discription;
     private String signature;
     private String annotation;
-    private String etype;
-    private Integer isMark;
 
-    public int getEpid() {
-        return epid;
+    public int getTeid() {
+        return teid;
     }
 
-    public void setEpid(int epid) {
-        this.epid = epid;
+    public void setTeid(int teid) {
+        this.teid = teid;
     }
 
     public String getDid() {
         return did;
     }
 
-    public void setDid(String mpid) {
+    public void setDid(String did) {
         this.did = did;
     }
 
-    public String getType() {
+    public Timestamp getTime() {
+        return time;
+    }
+
+    public void setTime(Timestamp time) {
+        this.time = time;
+    }
+
+    public int getEventtype() {
+        return eventtype;
+    }
+
+    public void setEventtype(int eventtype) {
+        this.eventtype = eventtype;
+    }
+
+    public int getType() {
         return type;
     }
 
-    public void setType(String type) {
+    public void setType(int type) {
         this.type = type;
     }
 
-    public String getContent() {
-        return content;
+    public int getSubtype() {
+        return subtype;
     }
 
-    public void setContent(String content) {
-        this.content = content;
+    public void setSubtype(int subtype) {
+        this.subtype = subtype;
     }
 
-    public Timestamp getOccurtime() {
-        return occurtime;
+    public int getDuration() {
+        return duration;
     }
 
-    public void setOccurtime(Timestamp occurtime) {
-        this.occurtime = occurtime;
+    public void setDuration(int duration) {
+        this.duration = duration;
+    }
+
+    public double getValue() {
+        return value;
+    }
+
+    public void setValue(double value) {
+        this.value = value;
+    }
+
+    public String getDiscription() {
+        return discription;
+    }
+
+    public void setDiscription(String discription) {
+        this.discription = discription;
     }
 
     public String getSignature() {
@@ -70,41 +104,27 @@ public class EventPower {
         this.annotation = annotation;
     }
 
-    public String getEtype() {
-        return etype;
-    }
-
-    public void setEtype(String etype) {
-        this.etype = etype;
-    }
-
-    public Integer getIsMark() {
-        return isMark;
-    }
-
-    public void setIsMark(Integer isMark) {
-        this.isMark = isMark;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         EventPower that = (EventPower) o;
-        return epid == that.epid &&
+        return teid == that.teid &&
+                eventtype == that.eventtype &&
+                type == that.type &&
+                subtype == that.subtype &&
+                duration == that.duration &&
+                Double.compare(that.value, value) == 0 &&
                 Objects.equals(did, that.did) &&
-                Objects.equals(type, that.type) &&
-                Objects.equals(content, that.content) &&
-                Objects.equals(occurtime, that.occurtime) &&
+                Objects.equals(time, that.time) &&
+                Objects.equals(discription, that.discription) &&
                 Objects.equals(signature, that.signature) &&
-                Objects.equals(annotation, that.annotation) &&
-                Objects.equals(etype, that.etype) &&
-                Objects.equals(isMark, that.isMark);
+                Objects.equals(annotation, that.annotation);
     }
 
     @Override
     public int hashCode() {
 
-        return Objects.hash(epid, did, type, content, occurtime, signature, annotation, etype, isMark);
+        return Objects.hash(teid, did, time, eventtype, type, subtype, duration, value, discription, signature, annotation);
     }
 }
