--- src/hibernatePOJO/EventTransient.hbm.xml
@@ -1,42 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-mapping PUBLIC
-    "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
-    "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
-<hibernate-mapping>
-
-    <class name="hibernatePOJO.EventTransient" table="event_transient" schema="test">
-        <id name="teid">
-            <column name="teid" sql-type="int(11)"/>
-        </id>
-        <property name="did">
-            <column name="did" sql-type="varchar(255)" not-null="true"/>
-        </property>
-        <property name="time">
-            <column name="time" sql-type="datetime"/>
-        </property>
-        <property name="eventtype">
-            <column name="eventtype" sql-type="int(11)"/>
-        </property>
-        <property name="type">
-            <column name="type" sql-type="int(11)"/>
-        </property>
-        <property name="subtype">
-            <column name="subtype" sql-type="int(11)"/>
-        </property>
-        <property name="duration">
-            <column name="duration" sql-type="int(11)"/>
-        </property>
-        <property name="value">
-            <column name="value" sql-type="double" precision="-1"/>
-        </property>
-        <property name="discription">
-            <column name="discription" sql-type="varchar(255)" not-null="true"/>
-        </property>
-        <property name="signature">
-            <column name="signature" sql-type="varchar(255)" not-null="true"/>
-        </property>
-        <property name="annotation">
-            <column name="annotation" sql-type="varchar(255)" not-null="true"/>
-        </property>
-    </class>
-</hibernate-mapping>
