+++ src/hibernatePOJO/EventsType.hbm.xml
@@ -1,0 +1,21 @@
+<?xml version='1.0' encoding='utf-8'?>
+<!DOCTYPE hibernate-mapping PUBLIC
+    "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
+    "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
+<hibernate-mapping>
+
+    <class name="hibernatePOJO.EventsType" table="events_type" schema="test">
+        <id name="type">
+            <column name="type" sql-type="int(11)"/>
+        </id>
+        <property name="name">
+            <column name="name" sql-type="varchar(255)" not-null="true"/>
+        </property>
+        <property name="subtypeset">
+            <column name="subtypeset" sql-type="varchar(255)" not-null="true"/>
+        </property>
+        <property name="prior">
+            <column name="prior" sql-type="int(11)" not-null="true"/>
+        </property>
+    </class>
+</hibernate-mapping>
