+++ src/hibernatePOJO/EventtypeDevice.hbm.xml
@@ -11,8 +11,5 @@
         <property name="name">
             <column name="name" sql-type="varchar(255)" not-null="true"/>
         </property>
-        <property name="priorty">
-            <column name="priorty" sql-type="varchar(255)" not-null="true"/>
-        </property>
     </class>
 </hibernate-mapping>
