+++ src/sms/getAlert.java
@@ -4,7 +4,6 @@ import Util.HBSessionDaoImpl;
 import com.alibaba.fastjson.JSONObject;
 import com.opensymphony.xwork2.ActionSupport;
 import hibernatePOJO.EventPower;
-import hibernatePOJO.EventTransient;
 import org.apache.struts2.ServletActionContext;
 
 import javax.servlet.http.HttpServletRequest;
@@ -44,11 +43,9 @@ public class getAlert extends ActionSupport {
             String date = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss:SS").format(calendar.getTime());
 
             //1.获得当前市行的15分钟内的告警事件（分为四类：）
-            List<EventTransient> alrmtranslist = hbsessionDao.search(
+            List<EventPower> alrmtranslist = hbsessionDao.search(
                     "FROM EventTransient where time >'"+ date +"'");
 
-            List<EventPower> alrmpowerlist = hbsessionDao.search(
-                    "FROM EventPower where occurtime >'"+ date +"' and isMark = 1");
 
             //2.查询各个事件的告警方式、告警用户、告警时间范围
             if(alrmtranslist != null){
@@ -78,9 +75,6 @@ public class getAlert extends ActionSupport {
                 }
             }
 
-            if(alrmpowerlist != null){
-
-            }
 
             System.out.println("完成告警(弹窗)模块:" + System.currentTimeMillis());
             JSONObject jsonObject = new JSONObject();
