+++ web/css/mycss.css
@@ -82,7 +82,7 @@ add-alarm-user{
     display: none;
     width: 500px;
     height: 440px;
-    background: rgba(0,0,0,.5);
+    background: rgba(0,0,0,.7);
     position: fixed;
     top: 0;
     left: 0;
@@ -99,7 +99,7 @@ add-alarm-user{
     display: block;
     width: 100px;
     height: 100px;
-    background: rgba(0,0,0,.5);
+    background: rgba(0,0,0,.7);
     position: fixed;
     top: 0;
     left: 0;
@@ -113,10 +113,10 @@ add-alarm-user{
 
 .conformEventRow-class {
     display: block;
-    width: 200px;
-    height: 150px;
+    width: 500px;
+    height: 400px;
     background: rgba(0,0,0,.7);
-    position: center;
+    position: fixed;
     top: 0;
     left: 0;
     right: 0;
